pbio/{control, servo}: Cast logged bool to int.
Fixes code smell.

+++ lib/pbio/src/control.c
@@ -308,7 +308,7 @@ void pbio_control_update(pbio_control_t *ctl, uint32_t time_now, pbio_control_st
             // Column 3: Speed in application units.
             pbio_control_settings_ctl_to_app(&ctl->settings, state->speed),
             // Column 4: Actuation type (LSB 0--1), stall state (LSB 2), on target (LSB 3), pause integration (LSB 4).
-            *actuation | (ctl->status << 2) | (pause_integration << 4),
+            *actuation | (ctl->status << 2) | ((int32_t)pause_integration << 4),
             // Column 5: Actuation payload, e.g. torque.
             *control,
             // Column 6: Reference position in application units.
